grpc_backend/requirements.txt
flask
flask-cors
pymongo
python-dotenv
bcrypt
PyJWT
grpcio
grpcio-tools
web3
